Read full desc, slight error handling change
As of right now, the ui doesnt "send the information anywhere", and I am not entirely sure how to go about doing that based on reading the other components code.

What needs to happen is that when the string is validated and added to the history list, it needs to pass that information to the lexer. Then if the lexer returns an error it should be caught by the ui and printed. Same in theory should happen if the lexer passes information to the parser.